// src/utils/dtos.ts
import { Category } from 'src/schemas/category.schema';

export class RetailerDto {
  firstName: string;
  lastName: string;
  email: string;
  password: string;
  status: string;
  phone: string;
  bName: string;
  shopDesc: string;
  logo: string;
  banner: string;
  industry: string;
  currency: string;
  abnOrAcn: string;
  yelloStoreRefId: string;
  ausPostStoreRefId: string;
  storeUrl: string;
  busDesc: string;
  ownerPhone: string;
  webAddress: string;
  isWebLive: boolean;
  country: string;
  add1: string;
  add2: string;
  suburb: string;
  state: string;
  postcode: string;
  avgorderval: string;
  skuCount: string;
  totalSale: string;
  pointOfSale: string;
  inventoryManagement: string;
  haveExistingAPi: boolean;
  transactionFees: string;
  expressShipFees: string;
  standardRangeforFree: string;
  intlRangeforFree: string;
  sendleRangeForFree: string;
  orderEmail: string;
  maxReturnDays: string;
  retailerSpecificDesc: string;
  bankAccName: string;
  bankName: string;
  bsb: string;
  accNum: string;
  accNumbtManager: string;
  sendleAPiKey: string;
  sendleId: string;
  sendlePlans: string;
  shipContry: string;
  affiliateLink: string;
  additionalInfo: string;
  isOnBoard: boolean;
  slug: string;
  categories: Category[];
}